Medical Infrastructure Management by Scheduling Visits and Online Access (chatbot)
The system “Medical Infrastructure Management by Scheduling Visits and Online Access” is a Machine Learning project which is basically used for taking inputs from user and provide a good results about their health. It is an interface that helps in providing appointment with specialist Doctor . it is an interface which is designed to give benefits to patients and users in a fast way as soon as possible.